refactor(library/HTTP): dissolve "/HTTPClient" into "/HTTP"
